+++ app/Species.php
@@ -15,8 +15,9 @@ class Species extends Model
     const WILDLIFE_USES = ['food', 'shelter'];
     const POLLINATING_TYPES = ['specialist', 'generalist'];
     const ECOLOGICAL_USES = ['riparian', 'slope', 'flood'];
+    const COMESTIBLE_USES = ['flower', 'fruit', 'leave', 'nut', 'seed', 'root'];
 
-    protected $fillable = ['name', 'name_en', 'name_fr', 'hardiness_ca', 'sun', 'soil', 'water', 'ph_min', 'ph_max', 'shape', 'root', 'maturity_height_meters', 'maturity_width_meters', 'nitrogen_fixer', 'nutrient_accumulator', 'ground_cover', 'hedge', 'wildlife_use', 'ecological_use', 'pollinating_type'];
+    protected $fillable = ['name', 'name_en', 'name_fr', 'hardiness_ca', 'sun', 'soil', 'water', 'ph_min', 'ph_max', 'shape', 'root', 'maturity_height_meters', 'maturity_width_meters', 'nitrogen_fixer', 'nutrient_accumulator', 'ground_cover', 'hedge', 'wildlife_use', 'ecological_use', 'pollinating_type', 'comestible_use', 'medicinal_use'];
 
     protected $casts = [
         'sun' => 'json',
@@ -26,6 +27,8 @@ class Species extends Model
         'wildlife_use' => 'json',
         'ecological_use' => 'json',
         'pollinating_type' => 'json',
+        'comestible_use' => 'json',
+        'medicinal_use' => 'boolean',
         'nitrogen_fixer' => 'boolean',
         'nutrient_accumulator' => 'boolean',
         'ground_cover' => 'boolean',