+++ database/migrations/2019_08_25_121507_alter_species_table_84.php
@@ -1,0 +1,33 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AlterSpeciesTable84 extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('species', function (Blueprint $table) {
+            $table->boolean('medicinal_use')->default(false);
+            $table->json('comestible_use')->nullable();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('species', function (Blueprint $table) {
+            $table->dropColumn(['medicinal_use', 'comestible_use']);
+        });
+    }
+}