+++ app/Http/Controllers/SpeciesController.php
@@ -4,13 +4,17 @@ namespace App\Http\Controllers;
 
 use App\Genus;
 use App\Rules\ComestibleUse;
+use App\Rules\Disadvantages;
 use App\Rules\EcologicalUse;
 use App\Rules\FloweringColor;
 use App\Rules\FloweringPeriod;
 use App\Rules\FoliageColor;
+use App\Rules\Growth;
 use App\Rules\HardinessCa;
+use App\Rules\Multiplication;
 use App\Rules\PollinatingType;
 use App\Rules\PostSummerAppeal;
+use App\Rules\PruningPeriod;
 use App\Rules\Root;
 use App\Rules\Shape;
 use App\Rules\Soil;
@@ -78,6 +82,10 @@ class SpeciesController extends Controller
             'flowering_color' => ['nullable', new FloweringColor],
             'foliage_color' => ['nullable', new FoliageColor],
             'post_summer_appeal' => ['nullable', new PostSummerAppeal],
+            'growth' => ['nullable', new Growth],
+            'pruning_period' => ['nullable', new PruningPeriod],
+            'multiplication' => ['nullable', new Multiplication],
+            'disadvantages' => ['nullable', new Disadvantages],
         ]);
 
         if ($validatedData['hardiness_ca'] == "null") {