+++ app/Species.php
@@ -20,8 +20,12 @@ class Species extends Model
     const FLOWERING_COLORS = ['red', 'rose', 'white', 'yellow', 'orange', 'purple'];
     const FOLIAGE_COLORS = ['gree', 'pale', 'dark', 'purple'];
     const POST_SUMMER_APPEALS = ['fall', 'winter'];
+    const GROWTH_SPEEDS = ['fast', 'medium', 'slow'];
+    const PRUNING_PERIODS = ['before budding', 'after flowering', 'summer', 'fall', 'anytime', "Don't prune"];
+    const MULTIPLICATIONS = ['cuttings', 'cuttings (spring)', 'cuttings (fall)', 'division', 'division (spring)', 'division (fall)', 'seedling', 'seedling (spring)', 'seedling (fall)', 'graft', 'graft (spring)', 'graft (fall)', 'stolon', 'stolon (spring)', 'stolon (fall)'];
+    const DISADVANTAGES = ['expansive', 'dispersive', 'allergen', 'poison', 'thorny'];
 
-    protected $fillable = ['name', 'name_en', 'name_fr', 'hardiness_ca', 'sun', 'soil', 'water', 'ph_min', 'ph_max', 'shape', 'root', 'maturity_height_meters', 'maturity_width_meters', 'nitrogen_fixer', 'nutrient_accumulator', 'ground_cover', 'hedge', 'wildlife_use', 'ecological_use', 'pollinating_type', 'comestible_use', 'medicinal_use', 'flowering_period', 'flowering_color', 'foliage_color', 'post_summer_appeal'];
+    protected $fillable = ['name', 'name_en', 'name_fr', 'hardiness_ca', 'sun', 'soil', 'water', 'ph_min', 'ph_max', 'shape', 'root', 'maturity_height_meters', 'maturity_width_meters', 'nitrogen_fixer', 'nutrient_accumulator', 'ground_cover', 'hedge', 'wildlife_use', 'ecological_use', 'pollinating_type', 'comestible_use', 'medicinal_use', 'flowering_period', 'flowering_color', 'foliage_color', 'post_summer_appeal', 'growth', 'pruning_period', 'multiplication', 'disadvantages'];
 
     protected $casts = [
         'sun' => 'json',
@@ -36,6 +40,10 @@ class Species extends Model
         'flowering_color' => 'json',
         'foliage_color' => 'json',
         'post_summer_appeal' => 'json',
+        'growth' => 'json',
+        'pruning_period' => 'json',
+        'multiplication' => 'json',
+        'disadvantages' => 'json',
         'medicinal_use' => 'boolean',
         'nitrogen_fixer' => 'boolean',
         'nutrient_accumulator' => 'boolean',