+++ database/migrations/2019_08_25_125222_alter_species_table_45.php
@@ -1,0 +1,35 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AlterSpeciesTable45 extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('species', function (Blueprint $table) {
+            $table->json('growth')->nullable();
+            $table->json('pruning_period')->nullable();
+            $table->json('multiplication')->nullable();
+            $table->json('disadvantages')->nullable();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('species', function (Blueprint $table) {
+            $table->dropColumn(['growth', 'pruning_period', 'multiplication', 'disadvantages']);
+        });
+    }
+}