+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 # 0.2.0 - Pre-release (2019-08-24)
 
-- Can add additional informations to _species_: architecture, functions.
+- Can add additional informations to _species_: architecture, functions, uses.
 
 # 0.1.0 - Pre-release (2019-08-24)